backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.22",
    "groq>=1.0.0",
]
